Responsive navbar for display users name.
